requirements.txt
-i https://pypi.org/simple
altgraph==0.17.3
async-generator==1.10 ; python_version >= '3.5'
asyncio==3.4.3
attrs==22.2.0 ; python_version >= '3.6'
beautifulsoup4==4.12.0
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1 ; os_name == 'nt' and implementation_name != 'pypy'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
customtkinter==5.1.2
darkdetect==0.8.0 ; python_version >= '3.6'
exceptiongroup==1.1.1 ; python_version < '3.11'
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
outcome==1.2.0 ; python_version >= '3.7'
pefile==2023.2.7 ; sys_platform == 'win32'
pillow==9.4.0
pycparser==2.21
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1 ; python_version >= '3.7'
pyotp==2.8.0
pyperclip==1.8.2
pysocks==1.7.1
pywin32-ctypes==0.2.0 ; sys_platform == 'win32'
requests==2.28.2
selenium==4.8.2
setuptools==67.6.0 ; python_version >= '3.7'
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
soupsieve==2.4 ; python_version >= '3.7'
tkinterweb==3.17.4
trio==0.22.0 ; python_version >= '3.7'
trio-websocket==0.10.2 ; python_version >= '3.7'
git+https://github.com/sebdelsol/undetected-chromedriver.git@1a337e41860f775607acec66d378929ffdffbb27#egg=undetected-chromedriver
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
websockets==10.4 ; python_version >= '3.7'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
